refactor: move code under seavision namespace package
