+++ seavision.egg-info/SOURCES.txt
@@ -1,36 +1,41 @@
 README.md
-pipeline.py
 pyproject.toml
-run_pipeline.py
-engine/__init__.py
-engine/postprocessor.py
-engine/detectors/__init__.py
-engine/detectors/base.py
-engine/detectors/motion/__init__.py
-engine/detectors/motion/background.py
-engine/detectors/motion/detector.py
-engine/detectors/motion/stabiliser.py
-engine/detectors/motion/tracker.py
-engine/detectors/sam3/__init__.py
-engine/detectors/sam3/config.py
-engine/detectors/sam3/detector.py
-engine/detectors/sam3/native.py
-engine/detectors/yolo/__init__.py
-engine/detectors/yolo/config.py
-engine/detectors/yolo/detector.py
-engine/source/__init__.py
-engine/source/base.py
-engine/source/discovery.py
-engine/source/local.py
-engine/source/s3.py
-engine/visualiser/__init__.py
-engine/visualiser/annotator.py
-engine/visualiser/config.py
-engine/visualiser/loader.py
-engine/visualiser/visualiser.py
-engine/visualiser/writer.py
+seavision/__init__.py
+seavision/pipeline.py
+seavision/run_pipeline.py
 seavision.egg-info/PKG-INFO
 seavision.egg-info/SOURCES.txt
 seavision.egg-info/dependency_links.txt
+seavision.egg-info/entry_points.txt
 seavision.egg-info/requires.txt
 seavision.egg-info/top_level.txt
+seavision/engine/__init__.py
+seavision/engine/postprocessor.py
+seavision/engine/detectors/__init__.py
+seavision/engine/detectors/base.py
+seavision/engine/detectors/motion/__init__.py
+seavision/engine/detectors/motion/background.py
+seavision/engine/detectors/motion/detector.py
+seavision/engine/detectors/motion/stabiliser.py
+seavision/engine/detectors/motion/tracker.py
+seavision/engine/detectors/sam3/__init__.py
+seavision/engine/detectors/sam3/config.py
+seavision/engine/detectors/sam3/detector.py
+seavision/engine/detectors/sam3/native.py
+seavision/engine/detectors/yolo/__init__.py
+seavision/engine/detectors/yolo/config.py
+seavision/engine/detectors/yolo/detector.py
+seavision/engine/source/__init__.py
+seavision/engine/source/base.py
+seavision/engine/source/discovery.py
+seavision/engine/source/local.py
+seavision/engine/source/s3.py
+seavision/engine/visualiser/__init__.py
+seavision/engine/visualiser/annotator.py
+seavision/engine/visualiser/config.py
+seavision/engine/visualiser/loader.py
+seavision/engine/visualiser/visualiser.py
+seavision/engine/visualiser/writer.py
+seavision/gui/__init__.py
+seavision/gui/app.py
+seavision/gui/validation/__init__.py
