+++ seavision.egg-info/requires.txt
@@ -1,3 +1,6 @@
 
 [dev]
 pytest
+
+[gui]
+PySide6>=6.5
