+++ seavision.egg-info/top_level.txt
@@ -1,3 +1,1 @@
-engine
-pipeline
-run_pipeline
+seavision
