+++ seavision/__init__.py
@@ -1,0 +1,3 @@
+"""SeaVision - Human-in-the-loop Computer Vision for Marine Environments."""
+
+__version__ = "0.1.0"
