+++ seavision/engine/detectors/motion/__init__.py
