+++ seavision/engine/detectors/motion/stabiliser.py
